PBI-VNT-HU2 | Codificar mejoras de funcionamiento
Se implementa la clase InfoVentas para cargar datos no comprometidos en una alta de venta, se implementa una solución para tratar los forms de altas y modificaciones como diferentes, se le da un formato a los campos de teléfono, no se permite modificar el número de empleado, se posicionan los MessageBoxConfirmar en el centro de la pantalla, se formatean los nombres de columnas en las tablas, se manejan decimales en vez de flotantes y se resuelves comportameintos extra`os de los formularios.

+++ carwash/negocios/VentasService.cs
@@ -13,7 +13,7 @@ namespace negocios
 
         private VentasDAO ventasDAO = new VentasDAO();
 
-        public void AltaVenta(string marcaCarro, string modeloCarro, string colorCarro, float precio, float gan, float corresp, int numEmp, DateTime fechaVenta) {
+        public void AltaVenta(string marcaCarro, string modeloCarro, string colorCarro, decimal precio, decimal gan, decimal corresp, int numEmp, DateTime fechaVenta) {
 
             ventasDAO.AltaVenta(marcaCarro, modeloCarro, colorCarro, precio, gan, corresp, numEmp, fechaVenta);
